test/requirements.txt
websockets
pyaudio
opuslib
fastapi
uvicorn[standard]
gunicorn
python-dotenv
google-genai>=1.70.0
httpx
requests
tzdata
openai
pytest
psycopg2
openpyxl